fix: Update some code according to TODOs
